Change quantidade type

+++ src/item_manager.c
@@ -74,7 +74,7 @@ static void set_item(GtkButton *b, GtkWidget *iman){
 		atof(gtk_entry_get_text(priv->icms_credito_aproveitado)),
 		item->imposto->icms);
 	item->valor = atof(gtk_entry_get_text(priv->valor));
-	item->quantidade = atof(gtk_entry_get_text(priv->quantidade));
+	item->quantidade = atoi(gtk_entry_get_text(priv->quantidade));
 	item->ordem = (ITEM_MANAGER(iman))->nfe->q_itens + 1; 
 	add_item(nfe, item);
 	gtk_widget_destroy(iman);
